requirements.txt
aiohttp
async-timeout
attrs
chardet
click
Flask
idna
itsdangerous
Jinja2
MarkupSafe
multidict
pycryptodomex
telepota
typing-extensions
urllib3
Werkzeug
yarl
youtube-dl